+++ Sources/SuperwallKit/Superwall.swift
@@ -127,6 +127,39 @@ public final class Superwall: NSObject, ObservableObject {
   @Published
   public var configurationStatus: ConfigurationStatus = .pending
 
+  /// A published property that indicates the subscription status of the user. If you're using a
+  /// `PurchaseController`, you must set this.
+  ///
+  /// If you're using Combine or SwiftUI, you can subscribe or bind to it to get
+  /// notified whenever it changes.
+  ///
+  /// Otherwise, you can check the delegate function
+  /// ``SuperwallDelegate/subscriptionStatusDidChange(from:to:)``
+  /// to receive a callback every time it changes.
+  @Published
+  public var subscriptionStatus: SubscriptionStatus = .unknown {
+    didSet {
+      if case let .active(entitlements) = subscriptionStatus {
+        if entitlements.isEmpty {
+          subscriptionStatus = .inactive
+          return
+        }
+      }
+      entitlements.subscriptionStatusDidSet(subscriptionStatus)
+    }
+  }
+
+  /// Returns the subscription status of the user.
+  ///
+  /// Check the delegate function
+  /// ``SuperwallDelegateObjc/subscriptionStatusDidChange(from:to:)``
+  /// to receive a callback every time it changes.
+  @available(swift, obsoleted: 1.0)
+  @objc(subscriptionStatus)
+  public var subscriptionStatusObjc: SubscriptionStatusObjc {
+    return subscriptionStatus.toObjc()
+  }
+
   /// The ``Entitlement``s tied to the device.
   public var entitlements: EntitlementsInfo {
     return dependencyContainer.entitlementsInfo
@@ -213,6 +246,9 @@ public final class Superwall: NSObject, ObservableObject {
     )
     self.init(dependencyContainer: dependencyContainer)
 
+    subscriptionStatus = dependencyContainer.storage.get(SubscriptionStatusKey.self) ?? .unknown
+    dependencyContainer.entitlementsInfo.subscriptionStatusDidSet(subscriptionStatus)
+
     addListeners()
 
     // This task runs on a background thread, even if called from a main thread.
@@ -270,6 +306,52 @@ public final class Superwall: NSObject, ObservableObject {
             }
           }
         ))
+
+    $subscriptionStatus
+      .removeDuplicates()
+      .dropFirst()
+      .scan((previous: subscriptionStatus, current: subscriptionStatus)) { previousPair, newStatus in
+        // Shift the current value to previous, and set the new status as the current value
+        (previous: previousPair.current, current: newStatus)
+      }
+      .receive(on: DispatchQueue.main)
+      .subscribe(Subscribers.Sink(
+        receiveCompletion: { _ in },
+        receiveValue: { [weak self] statusPair in
+          guard let self = self else {
+            return
+          }
+          let oldStatus = statusPair.previous
+          let newStatus = statusPair.current
+
+          self.dependencyContainer.storage.save(newStatus, forType: SubscriptionStatusKey.self)
+
+          Task {
+            await self.dependencyContainer.delegateAdapter.subscriptionStatusDidChange(from: oldStatus, to: newStatus)
+            let event = InternalSuperwallPlacement.SubscriptionStatusDidChange(status: newStatus)
+            await Superwall.shared.track(event)
+          }
+        }
+      ))
+  }
+
+  /// Sets ``subscriptionStatus`` to an`unknown` state.
+  @available(swift, obsoleted: 1.0)
+  public func setUnknownSubscriptionStatus() {
+    subscriptionStatus = .unknown
+  }
+
+  /// Sets ``subscriptionStatus`` to an`inactive` state.
+  @available(swift, obsoleted: 1.0)
+  public func setInactiveSubscriptionStatus() {
+    subscriptionStatus = .inactive
+  }
+
+  /// Sets ``subscriptionStatus`` to an`active` state with the
+  /// specified entitlements.
+  @available(swift, obsoleted: 1.0)
+  public func setActiveSubscriptionStatus(with entitlements: Set<Entitlement>) {
+    subscriptionStatus = .active(entitlements)
   }
 
   // MARK: - Configuration
@@ -284,7 +366,7 @@ public final class Superwall: NSObject, ObservableObject {
   ///   an account, you can [sign up for free](https://superwall.com/sign-up).
   ///   - purchaseController: An optional object that conforms to ``PurchaseController``. Implement this if you'd
   ///   like to handle all subscription-related logic yourself. You'll need to also set the ``subscriptionStatus`` every time the user's
-  ///   entitlements change. You can read more about that in [Purchases and Entitlements](https://docs.superwall.com/docs/advanced-configuration).
+  ///   entitlements change. You can read more about that in [Purchases and Subscription Status](https://docs.superwall.com/docs/advanced-configuration).
   ///   If `nil`, Superwall will handle all subscription-related logic itself.  Defaults to `nil`.
   ///   - options: An optional ``SuperwallOptions`` object which allows you to customise the appearance and behavior
   ///   of the paywall.
@@ -334,8 +416,8 @@ public final class Superwall: NSObject, ObservableObject {
   ///   - apiKey: Your Public API Key that you can get from the Superwall dashboard settings. If you don't have an account, you
   ///   can [sign up for free](https://superwall.com/sign-up).
   ///   - purchaseController: An optional object that conforms to ``PurchaseControllerObjc``. Implement this if you'd
-  ///   like to handle all subscription-related logic yourself. You'll need to also set the ``entitlements`` every time the user's
-  ///   entitlements change. You can read more about that in [Purchases and Entitlements](https://docs.superwall.com/docs/advanced-configuration).
+  ///   like to handle all subscription-related logic yourself. You'll need to also set the ``subscriptionStatus`` every time the user's
+  ///   entitlements change. You can read more about that in [Purchases and Subscription Status](https://docs.superwall.com/docs/advanced-configuration).
   ///   If `nil`, Superwall will handle all subscription-related logic itself.  Defaults to `nil`.
   ///   - options: A ``SuperwallOptions`` object which allows you to customise the appearance and behavior of the paywall.
   ///   - completion: An optional completion handler that lets you know when Superwall has finished configuring.