+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
+## 4.0.0-beta.5
+
+### Breaking Changes
+
+- Replaces `entitlements.status` with `subscriptionStatus`.
+
+### Enhancements
+
+- Adds `Superwall.shared.subscriptionStatus.isActive` as a convenience variable.
+
 ## 4.0.0-beta.4
 
 ### Fixes